"A lot has been created"

+++ Entities/DbSets/LicenseClass.cs
@@ -7,7 +7,7 @@ public class LicenseClass
     public int Id { get; set; }
     public string ClassName { get; set; }
     public string ClassDescription { get; set; }
-    public byte MinimumAllowedAge { get; set; }
+    public int MinimumAllowedAge { get; set; }
     public int DefaultValidityLength { get; set; }
     public float ClassFees { get; set; }
     public ICollection<License>? Licenses { get; set; }
